+++ src/instagram-feed/instagram-feed.js
@@ -34,7 +34,7 @@ function InstagramFeed({ slide, content, run, slideDone, executionId }) {
   const animationDuration = 1500;
 
   const { feedData = [] } = slide;
-  const { hashtagText, orientation, imageWidth = null } = content;
+  const { hashtagText, orientation, imageWidth = null, mediaContain } = content;
 
   // @TODO: should duration depend on number of instagram posts to show?
   let { entryDuration: duration } = content;
@@ -104,7 +104,7 @@ function InstagramFeed({ slide, content, run, slideDone, executionId }) {
           <div className="media-section">
             {!currentPost.videoUrl && (
               <div
-                className="image"
+                className={`image${mediaContain ? " media-contain" : ""}`}
                 style={{
                   backgroundImage: `url("${currentPost.mediaUrl}")`,
                   ...(show
@@ -115,7 +115,13 @@ function InstagramFeed({ slide, content, run, slideDone, executionId }) {
             )}
             {currentPost.videoUrl && (
               <div className="video-container">
-                <video muted="muted" autoPlay loop src={currentPost.videoUrl}>
+                <video
+                  muted="muted"
+                  autoPlay
+                  loop
+                  src={currentPost.videoUrl}
+                  className={mediaContain ? "media-contain" : ""}
+                >
                   <track kind="captions" />
                 </video>
               </div>
@@ -169,6 +175,7 @@ InstagramFeed.propTypes = {
     entryDuration: PropTypes.number,
     maxEntries: PropTypes.number,
     imageWidth: PropTypes.number,
+    mediaContain: PropTypes.bool,
   }).isRequired,
   executionId: PropTypes.string.isRequired,
 };